requirements.txt
pytesseract
pillow
langdetect
